Cleaned up the code and added examples.
