+++ src/cli/repl.ts
@@ -3,7 +3,7 @@ import { createInterface } from 'readline';
 import chalk from 'chalk';
 import { Swarm } from '../core/swarm';
 import OpenAI from 'openai';
-import DagSwarm from '../lib/swarms/DagSwarm';
+import { Agent } from '../core/types';
 
 const DEBUG = process.env.DEBUG === 'true';
 
@@ -11,59 +11,6 @@ const rl = createInterface({
     input: process.stdin,
     output: process.stdout,
 });
-
-async function create_greeting(names: string, timeOfDay: string): Promise<string> {
-    return `Good ${timeOfDay}, ${names}!`;
-}
-
-async function get_time_of_day(timeZone: string | { timeZone: string } = '+00:00'): Promise<string> {
-    const date = new Date();
-    let hour: number;
-    
-    const tzString = typeof timeZone === 'object' ? timeZone.timeZone : timeZone;
-    const timezoneMap: { [key: string]: string } = {
-        'PST': 'America/Los_Angeles',
-        'EST': 'America/New_York',
-        'MST': 'America/Denver',
-        'CST': 'America/Chicago',
-    };
-
-    const ianaTimezone = timezoneMap[tzString.toUpperCase()] || tzString;
-
-    if (ianaTimezone.includes('/')) {
-        try {
-            const options: Intl.DateTimeFormatOptions = { 
-                timeZone: ianaTimezone, 
-                hour: "numeric" as const, 
-                hour12: false 
-            };
-            hour = parseInt(new Intl.DateTimeFormat('en-US', options).format(date));
-        } catch (error) {
-            throw new Error("Invalid IANA timezone name");
-        }
-    } else {
-        const tzParts = tzString.match(/^([+-])(\d{2}):?(\d{2})?$/);
-        if (!tzParts) {
-            throw new Error("Invalid timezone format. Expected +HH:MM or -HH:MM");
-        }
-        
-        const sign = tzParts[1] === '+' ? -1 : 1;
-        const hours = parseInt(tzParts[2]);
-        const minutes = tzParts[3] ? parseInt(tzParts[3]) : 0;
-        const requestedOffset = sign * (hours * 60 + minutes);
-        
-        const offsetDiff = requestedOffset + date.getTimezoneOffset();
-        hour = (date.getHours() + Math.floor(offsetDiff / 60)) % 24;
-        if (hour < 0) hour += 24;
-    }
-
-    if (hour < 12) return "morning";
-    else if (hour < 17) return "afternoon";
-    else return "evening";
-}
-
-create_greeting.description = 'Given a person\'s name, return a greeting message.';
-get_time_of_day.description = 'Get the time of day (morning, afternoon, or evening) for the given timezone.';
 
 async function processAndPrintStreamingResponse(response: AsyncGenerator<any, void, unknown>) {
     let content = '';
@@ -98,23 +45,21 @@ async function processAndPrintStreamingResponse(response: AsyncGenerator<any, vo
     }
 }
 
-async function runDemoLoop() {
+export async function runExample(
+    name: string,
+    getAgent: () => Promise<Agent> | Agent
+) {
     const client = new OpenAI({
         apiKey: process.env.OPENAI_API_KEY,
     });
 
     const swarm = new Swarm(client);
-    const functions = [create_greeting, get_time_of_day];
     
-    const dagSwarm = new DagSwarm(
-        "create a customized greeting message for the given name and the timezone that the user provided",
-        functions
-    );
-
-    console.log(chalk.green('Starting SwarmJS CLI with DagSwarm 🐝'));
+    console.log(chalk.green(`Starting SwarmJS CLI with ${name} 🐝`));
     console.log(chalk.gray('Type your messages and press Enter. Press Ctrl+C to exit.\n'));
 
     const messages: any[] = [];
+    const agent = await Promise.resolve(getAgent());
 
     while (true) {
         const userInput = await new Promise<string>((resolve) => {
@@ -131,7 +76,7 @@ async function runDemoLoop() {
 
         try {
             const response = await swarm.run(
-                dagSwarm.getDagCreationAgent(),
+                agent,
                 messages,
                 {},  // context variables
                 null, // model override
@@ -151,13 +96,13 @@ async function runDemoLoop() {
     }
 }
 
+if (require.main === module) {
+    console.log(chalk.yellow('Please run a specific example file instead of this generic REPL.'));
+    process.exit(1);
+}
+
 process.on('SIGINT', () => {
     console.log(chalk.yellow('\nGoodbye! 👋'));
     rl.close();
     process.exit(0);
-});
-
-runDemoLoop().catch((error) => {
-    console.error(chalk.red('Error:'), error);
-    process.exit(1);
 });
